==> Step1_Translate.ipynb <==
# Abstract

<p>This notebook processes translates all of the BAM files aligned to Tat exons 1 and 2. All of the translated files are saved in /FullNeuroIllumina/Counts/Tat1/ or /FullNeuroIllumina/Counts/Tat2/, depending on the exon being processed, as a .tsv file.</p>
<p>Runing this notebook can take a long time because the translation steps are fairly slow.</p>
<p>All of the analysis for this notebook can be considered complete<p>

# Requirements

from __future__ import division
import sys
import pysam
import glob
import pandas as pd
import numpy as np
import itertools
from collections import Counter
from Bio.Seq import Seq
from Bio.Alphabet import generic_dna
import time
from collections import defaultdict
import csv
from itertools import islice
import matplotlib.pyplot as plt
%matplotlib inline

sys.path.append('/Users/greg/damlarces/')
from damlarces.NGSTools import ReadTools

files = glob.glob('/Users/greg/Desktop/FullNeuroIllumina/BAMs/*.bam')
f = files[0]
print len(files)
print f

#TESTING AND CHECKING
out = '/Users/greg/Desktop/tsvcheck.A0001-R09.tsv'
print out
samfile = pysam.Samfile(f, mode='rb').fetch(region='HXB2:5800-6060')
counts = defaultdict(int)
it = ReadTools.SeqReadRecord.translate_from_stream(samfile, [('Tat1', 5830, 6045)])
for orf in it:
    for (read, ref), l in zip(orf.aligned_pairs, orf.seq):
        if l != '-':
            counts[(orf.chrom, ref, l)] += 1
        if l == '*':
            break
with open(out, 'w') as handle:
    writer = csv.writer(handle, delimiter = '\t')
    writer.writerow(('Prot', 'AAPos', 'AA', 'Count'))
    for (prot, pos, l), n in sorted(counts.items()):
        writer.writerow((prot, pos, l, n))

ReadTools.HXB2_ORFS[3:-1]

for f in files:
    out = f.replace('hivsorted.bam', 'Tat1.AAcounts.tsv').replace('/BAMs/', '/Counts/Tat1/')
    print out
    samfile = pysam.Samfile(f, mode='rb').fetch(region='HXB2:5800-6060')
    counts = defaultdict(int)
    it = ReadTools.SeqReadRecord.translate_from_stream(samfile, [('Tat1', 5830, 6044)])
    for orf in it:
        for (read, ref), l in zip(orf.aligned_pairs, orf.seq):
            if l != '-':
                counts[(orf.chrom, ref, l)] += 1
            if l == '*':
                break
    with open(out, 'w') as handle:
        writer = csv.writer(handle, delimiter = '\t')
        writer.writerow(('Prot', 'AAPos', 'AA', 'Count'))
        for (prot, pos, l), n in sorted(counts.items()):
            writer.writerow((prot, pos, l, n))

for f in files:
    out = f.replace('hivsorted.bam', 'Tat2.AAcounts.tsv').replace('/BAMs/', '/Counts/Tat2/')
    print out
    samfile = pysam.Samfile(f, mode='rb').fetch(region='HXB2:8370-8480')
    counts = defaultdict(int)
    it = ReadTools.SeqReadRecord.translate_from_stream(samfile, [('Tat2', 8378, 8468)])
    for orf in it:
        for (read, ref), l in zip(orf.aligned_pairs, orf.seq):
            if l != '-':
                counts[(orf.chrom, ref, l)] += 1
            if l == '*':
                break
    with open(out, 'w') as handle:
        writer = csv.writer(handle, delimiter = '\t')
        writer.writerow(('Prot', 'AAPos', 'AA', 'Count'))
        for (prot, pos, l), n in sorted(counts.items()):
            writer.writerow((prot, pos, l, n))
