==> Step2B_ConsensusAndReconstructions.ipynb <==
# Abstract

<p>This notebook evalutes the consensus sequence for each patient using NGS data.</p>
<p>It also reads in the full Tat reconstructions.</p>

# Requirements

from __future__ import division
import glob
import pandas as pd
import numpy as np
from collections import Counter
import matplotlib.pyplot as plt
%matplotlib inline

# Consensus

freq_files = glob.glob('/Users/greg/Desktop/FullNeuroIllumina/Frequency/Tat1/*AAfreqcount.csv')

cols = ['A','R','N','D','C','Q','E','G','H','I',
        'L','K','M','F','P','S','T','W','Y','V']

for freq_file in freq_files:
    df = pd.read_csv(freq_file)
    if df.shape[0] > 70:
        print df.shape
        a = df[cols].idxmax(axis=1, skipna=True)
        print freq_file.split('/')[-1].split('.')[0]
        print ''.join(list(a))

freq_files = glob.glob('/Users/greg/Desktop/FullNeuroIllumina/Frequency/Tat2/*AAfreqcount.csv')

cols = ['A','R','N','D','C','Q','E','G','H','I',
        'L','K','M','F','P','S','T','W','Y','V']

for freq_file in freq_files:
    df = pd.read_csv(freq_file)
    if df.shape[0] > 12:
        a = df[cols].idxmax(axis=1, skipna=True)
        print freq_file.split('/')[-1].split('.')[0]
        print ''.join(list(a))
