==> Step4_LoadClinical.ipynb <==
# Abstract

<p>This notebook loads all the redcap data and full neuro data, and merges these two data sets appropriately.</p>
<p>The result of this notebook is a dataframe containing the full extent of clinical data for the Drexel CARES cohort.</p>
<p>This data is saved at /Users/greg/Desktop/FullNeuroIllumina/Clinical/full_clinical.csv</p>

# Import requirements

from __future__ import division
import glob
import pandas as pd
import numpy as np
import itertools
from collections import Counter
import time
import matplotlib.pyplot as plt
%matplotlib inline

# Load redcap data

def import_redcap_data(redcap_path):
    redcap_df = pd.read_csv(redcap_path, sep='\t')
    redcap_df.head()
    return redcap_df

def select_clinical_parameters(df, cols):
    clin_df = df[cols.keys()].rename(columns=cols)
    clin_df['DateOfVisit'] = pd.to_datetime(clin_df['DateOfVisit'])
    return clin_df

redcap_path = '/Users/greg/Desktop/Cohort/DrexelMedCARESCohort_DATA_LABELS_2015-10-06_1434.csv'
cols = {'PatientID':'Patient',
        'Patient visit number':'Visit',
        'Date of visit':'DateOfVisit',
        'Initial CD4 count (cells/uL)':'iCD4',
        'Nadir CD4 count (cells/uL)':'nCD4',
        'Latest CD4 count (cells/uL)':'CD4',
        'Initial CD8 count (cells/uL)': 'iCD8',
        'Nadir CD8 count (cells/uL)': 'nCD8',
        'Latest CD8 count (cells/uL)': 'CD8',
        'Initial viral load (copies/mL)': 'iVL',
        'Peak viral load (copies/mL)': 'pVL',
        'Latest viral load': 'VL',
        'Total Modified Hopkins Dementia Score': 'TMHDS',
        'Gender': 'Gender',
        'Age': 'Age',
        'Current ART status': 'ART'}

col_order = ['Patient', 'Visit', 'DateOfVisit', 'Age', 'Gender', 'ART',
             'VL', 'iVL', 'pVL', 'CD4', 'iCD4', 'nCD4', 'CD8', 'iCD8', 'nCD8',
             'TMHDS']


redcap_df = import_redcap_data(redcap_path)
redcap_df['PatientID'] = redcap_df['PatientID'].astype(str)
print 'REDCAP:', redcap_df.shape

clin_df = select_clinical_parameters(redcap_df, cols)
clin_df = clin_df[col_order]
print 'CLIN:', clin_df.shape
clin_df.head(5)

# Load GDS data

gds_path = '/Users/greg/Desktop/Cohort/Drexel HIV NP Data select standardized 072815 - fixed.xlsx'
GDS_df = pd.read_excel(gds_path, headline=1)
GDS_df['PatientID'] = GDS_df['PatientID'].astype(str)
print 'GDS:', GDS_df.shape
GDS_df.head()

# GDS visit corrections

visits = []
for idx, row in GDS_df.iterrows():
    NEURO_PATIENT = row['PatientID']
    NEURO_DATE = row['VisitDate']    
    
    clin_info = clin_df[clin_df.Patient == NEURO_PATIENT][['Patient', 'Visit', 'DateOfVisit']]
    clin_info2 = clin_info[clin_info.DateOfVisit == NEURO_DATE][['Patient', 'Visit', 'DateOfVisit']]
    if clin_info2.empty:
        #print NEURO_PATIENT, NEURO_DATE
        for a, b in clin_info.dropna().iterrows():
            days_difference = b.DateOfVisit - NEURO_DATE
            if abs(days_difference.days) < 7:
                visits.append(b.Visit)
                #print ''
    else:
        #print NEURO_PATIENT, NEURO_DATE
        visits.append(list(clin_info2['Visit'])[0])
        #print ''
        
GDS_df['Visit'] = visits
print GDS_df.shape
GDS_df[['PatientID', 'Visit', 'VisitDate', 'GDS']].head()

# Merge Clinical information with full neuro data

fullclinical_df = pd.merge(clin_df,GDS_df,
                  left_on = ['Patient','Visit'],
                  right_on = ['PatientID','Visit'],
                  how = 'outer')

fullclinical_df.drop(['PatientID'], axis=1, inplace=True)
print fullclinical_df.shape
fullclinical_df.head()

# Numerical value corrections

incorrect_columns = ['iCD4','nCD4','CD4','iCD8','nCD8','CD8','iVL','pVL','VL']

for colname in incorrect_columns:
    #print colname
    corrected_values = []
    for item in list(fullclinical_df[colname]):
        l = []
        if isinstance(item, float):
            corrected_values.append(item)
        elif isinstance(item, str):
            item = item.replace('<','')
            item = item.replace('>','')
            item = item.replace(',','')
            item = item.replace('?','')
            for t in item.split():
                try:
                    l.append(float(t))
                except ValueError:
                    #pass
                    l.append(np.nan)
                except IndexError:
                    l.append(np.nan)
            #print l, item
            corrected_values.append(l[0])
        else:
            print item, type(item)
    #print len(corrected_values)
    fullclinical_df[colname] = corrected_values

print 'Full Clinical:', fullclinical_df.shape
fullclinical_df.head()

fullclinical_df.to_csv('/Users/greg/Desktop/FullNeuroIllumina/Clinical/full_clinical.csv', index=False)
