==> Step5_Merge_NGS_and_Clinical.ipynb <==
# Abstract

<p>This notebook reads in the NGS data (all 101 positions of Tat) as well as the full clinical information containing all of the neuropsych scoring metrics</p>

# Import requirements

from __future__ import division
import glob
import pandas as pd
import numpy as np
import itertools
from collections import Counter
import time
import matplotlib.pyplot as plt
%matplotlib inline

# Import sequence data

# NGS proportional abundance
seq_abundance_df = pd.read_csv('/Users/greg/Desktop/FullNeuroIllumina/NGSProfiles/NGSprofile.csv')
print seq_abundance_df.shape
seq_abundance_df.head()

# NGS raw counts
seq_counts_df = pd.read_csv('/Users/greg/Desktop/FullNeuroIllumina/NGSProfiles/NGScounts.csv')
print seq_counts_df.shape
seq_counts_df.head()

# Import clinical data

clinical_df = pd.read_csv('/Users/greg/Desktop/FullNeuroIllumina/Clinical/full_clinical.csv')
print clinical_df.shape
clinical_df.head()

# Merge NGS and Clinical data

# relative proportion data
merged_df1 = pd.merge(clinical_df, seq_abundance_df,
                  left_on = ['Patient','Visit'],
                  right_on = ['Patient','Visit'],
                  how = 'inner')
merged_df1.sort_values(['AAPos','Patient','Visit'],  inplace=True)
print merged_df1.shape

# sequence counts data
merged_df2 = pd.merge(clinical_df, seq_counts_df,
                  left_on = ['Patient','Visit'],
                  right_on = ['Patient','Visit'],
                  how = 'inner')
merged_df2.sort_values(['AAPos','Patient','Visit'],  inplace=True)
print merged_df2.shape

GDSmerged_df = merged_df1.dropna(axis=0, subset=['GDS'])
TMHDSmerged_df = merged_df1.dropna(axis=0, subset=['TMHDS'])
TMHDSGDSmerged_df = TMHDSmerged_df.dropna(axis=0, subset=['GDS'])
print GDSmerged_df.shape
print TMHDSmerged_df.shape
print TMHDSGDSmerged_df.shape

GDSmerged_counts_df = merged_df2.dropna(axis=0, subset=['GDS'])
TMHDSmerged_counts_df = merged_df2.dropna(axis=0, subset=['TMHDS'])
TMHDSGDSmerged_counts_df = TMHDSmerged_counts_df.dropna(axis=0, subset=['GDS'])
print GDSmerged_counts_df.shape
print TMHDSmerged_counts_df.shape
print TMHDSGDSmerged_counts_df.shape

# Merge dataframes

# merged proportional data
GDSmerged_df.to_csv('/Users/greg/Desktop/FullNeuroIllumina/MergedData/NGS_GDS_abundance.csv', index=False)
TMHDSmerged_df.to_csv('/Users/greg/Desktop/FullNeuroIllumina/MergedData/NGS_TMHDS_abundance.csv', index=False)
TMHDSGDSmerged_df.to_csv('/Users/greg/Desktop/FullNeuroIllumina/MergedData/NGS_BOTH_abundance.csv', index=False)

# merged raw counts data
GDSmerged_counts_df.to_csv('/Users/greg/Desktop/FullNeuroIllumina/MergedData/NGS_GDS_counts.csv', index=False)
TMHDSmerged_counts_df.to_csv('/Users/greg/Desktop/FullNeuroIllumina/MergedData/NGS_TMHDS_counts.csv', index=False)
TMHDSGDSmerged_counts_df.to_csv('/Users/greg/Desktop/FullNeuroIllumina/MergedData/NGS_BOTH_counts.csv', index=False)

#GDS_merged_df.to_csv('/Users/greg/Desktop/FullNeuroIllumina/GDS_only_merged.csv', index=False)
#merged_df.to_csv('/Users/greg/Desktop/FullNeuroIllumina/merged_data.csv', index=False)

# Survey data

def slice_position(df, pos):
    m = df['AAPos'] == pos
    sliced_df = df[m]
    return sliced_df

def tally_groups(dataframe, parameter, threshold):
    df = slice_position(dataframe, 1)
    a = sum(df[parameter] < threshold)
    b = sum(df[parameter] >= threshold)
    return (a,b)

DFs = [GDSmerged_df, TMHDSmerged_df, TMHDSGDSmerged_df,
       GDSmerged_counts_df, TMHDSmerged_counts_df,
       TMHDSGDSmerged_counts_df]

for i in range(1,72+1):
    for g in DFs:
        (a,b) = tally_groups(g, 'GDS', 0.5)
        print((a,b)),
    print '\n'
