==> Step7_statsmodels_LogisticRegression.ipynb <==
# Import Requirements

from __future__ import division
import pandas as pd
import numpy as np
from collections import Counter
import time
import statsmodels.api as sm
import matplotlib.pyplot as plt
%matplotlib inline

# Import data

sample_col = ['Patient','Visit']
demographics_col = ['Age', 'Gender', 'ART']
chemistry_col = ['VL','iVL','pVL','CD4','iCD4','nCD4','CD8','iCD8',u'nCD8']
neuro_col =['TMHDS','GDS']
sequence_col = ['Prot','AAPos','Coverage']
aa_col = ['A','R','N','D','C','Q','E','G','H','I','L','K','M','F','P','S','T','W','Y','V']

DF = pd.read_csv('/Users/greg/Desktop/FullNeuroIllumina/merged_data.csv')
DF.head()

# Define functions

def slice_position(df, pos):
    m = df['AAPos'] == pos
    sliced_df = df[m]
    return sliced_df

def chooseDataCols(pos_df):
    cols = []
    cols.extend(sample_col)
    cols.extend(neuro_col)
    cols.extend(aa_col)
    data = pos_df[cols].copy()
    data['intercept'] = 1.0
    return data

def filterX(X, threshold):
    t = X.max()>threshold
    #t = X.mean()>threshold
    filtered_cols = t[t==True].index.tolist()
    X = X[filtered_cols]
    return X

def getXandY(data, threshold):   
    train_cols = []
    train_cols.extend(aa_col)
    train_cols.extend(['intercept'])
    y = (data['GDS'] >= 0.5).astype(int)
    X = data[train_cols]
    X = filterX(X, threshold)
    return X, y

# One position example

pos_df = slice_position(DF,61)
data = chooseDataCols(pos_df)
X, y = getXandY(data, 0.05)
X = X*100
X['intercept'] = 1
# initialize and fit the model
logit = sm.Logit(y, X)
result = logit.fit()
print result.summary()
#print result.params
#print result.conf_int()
#for i,row in result.conf_int().iterrows():
#    print i,row[0],row[1]

# All positions

tat1_length=72
for i in range(1,tat1_length+1):
    #slice the data
    pos_df = slice_position(DF, i)
    
    # prep the data
    data = chooseDataCols(pos_df)
    X, y = getXandY(data,0.05)
    print (i, X.shape, y.shape)

    # initialize and fit the model
    try:
        logit = sm.Logit(y, X)
        result = logit.fit()
        print result.conf_int()
    except:
        continue
    print ''
