==> Step9_ConsensusApproach.ipynb <==
# Import requirements

from __future__ import division
import glob
import math
import pandas as pd
import numpy as np
from itertools import combinations
from collections import Counter
import time
import matplotlib.pyplot as plt
%matplotlib inline

# Import Data

DF = pd.read_csv('/Users/greg/Desktop/FullNeuroIllumina/merged_data.csv')
print DF.shape
DF.head()

# Functions

def GDS_split(df, parameter, threshold):
    mask1 = df[parameter] < threshold
    mask2 = df[parameter] >= threshold
    df1 = df[mask1]
    df2 = df[mask2]
    return df1, df2

def reformat_section(df):
    '''Must be a dataframe containing only 1 position'''
    pos = str(int(list(df.AAPos)[0]))
    AAs = ['A','R','N','D','C','Q','E','G','H','I',
           'L','K','M','F','P','S','T','W','Y','V']
    newnames_dict = {}
    for item in AAs:
        newnames_dict[item]= pos+item
    df_renamed = df.rename(columns = newnames_dict)
    df_renamed.drop(['AAPos','Coverage'], axis=1, inplace=True)
    return df_renamed

def consensusCallsDF(df):
    consensus_dict = {'Position':[], 'Calls':[], 'Richness':[]}
    for i,group in df.groupby('AAPos'):
        calls = []
        for j, sample in group.iterrows():
            if max(sample[AAs]) > 0:
                sample_call = str(sample[AAs].idxmax())
                #sample_GDS = sample['GDS']
                calls.append(sample_call)
        #print int(i), dict(Counter(calls))
        consensus_dict['Position'].append(int(i))
        consensus_dict['Calls'].append(dict(Counter(calls)))
        consensus_dict['Richness'].append(len(Counter(calls)))
    consensusDF = pd.DataFrame(consensus_dict)
    return consensusDF

# Split into groups by GDS

DF1, DF2 = GDS_split(DF, 'GDS', 0.5)
print DF1.shape
print DF2.shape

conDF = consensusCallsDF(DF)
conDF1 = consensusCallsDF(DF1)
conDF2 = consensusCallsDF(DF2)
print conDF.shape
print conDF1.shape
print conDF2.shape

pd.DataFrame(zip(conDF.Position, conDF.Richness, conDF1.Richness, conDF2.Richness)).head()

# Consensus "Reconstruction"

DF.sort_values(by=['Patient','Visit']).head()

d = {'Patient':[], 'Visit':[], 'GDS Sequence':[]}
for i, group in DF.groupby(['Patient','Visit']):
    s = ''
    for pos in range(1,73):  
        g = group[group['AAPos']==pos][AAs]
        if g.shape[0]:
            aa = g.transpose().idxmax().as_matrix()[0]
        else:
            aa = '-'
        s+=aa
    d['Patient'].append(i[0])
    d['Visit'].append(i[1])
    d['GDS Sequence'].append(s)
reconstructionDF = pd.DataFrame(d)
print reconstructionDF.shape
reconstructionDF.head()

# Sanger data

sanger_df = pd.read_csv('/Users/greg/Desktop/TMHDSNeuroTat/all_Tat_data_df_2015_12_10.csv')
sanger_df.head()

# Merge Illumina and Sanger sequence data

seqcompare_df = pd.merge(reconstructionDF, sanger_df,
                left_on=['Patient','Visit'],
                right_on=['PatientID','VisitNum'],
                how='inner')

seqcompare_df.head()

# Compare Illumina and Sanger sequences for matching samples

# make dataframe

d = {'Patient':[], 'Visit':[], 'Illumina':[], 'Sanger':[]}
for i, group in seqcompare_df.groupby(['Patient','Visit']):
    #print (i, group.shape)
    s_sanger = ''
    for j in np.arange(1,73):
        position_col = 'Pos%02i' % j
        aa = list(group[position_col])[0]
        if isinstance(aa, float):
            if math.isnan(aa):
                aa = '-'
        s_sanger+=aa
    s_illumina = list(group['GDS Sequence'])[0]
    
    denom = 0
    num = 0
    for pair in zip(s_sanger, s_illumina):
        if '-' not in pair:
            denom+=1
            if pair[0]==pair[1]:
                num+=1
    if denom>0:
        agreement = num/denom
        d['Patient'].append(i[0])
        d['Visit'].append(i[1])
        d['Sanger'].append(s_sanger)
        d['Illumina'].append(s_illumina)
colorder = ['Patient','Visit','Sanger','Illumina']
comp_df = pd.DataFrame(d)[colorder]
comp_df.head()

# Agreement function

def calcSeqAgreement(seq1, seq2):
    denom = 0
    num = 0
    for pair in zip(seq1, seq2):
        if '-' not in pair:
            denom+=1
        if pair[0]==pair[1]:
            num+=1
    if denom>0:
        agreement = num/denom
    else:
        agreement = 0
    return (denom, num, agreement)

# Calculate agreement

sanger_seqs = list(comp_df['Sanger'])
illumina_seqs = list(comp_df['Illumina'])
z = zip(sanger_seqs, illumina_seqs)
scores = []
for pair in z:
    a = calcSeqAgreement(pair[0], pair[1])
    scores.append(a[2])
print np.mean(scores)
print np.std(scores)

scores = []
for combo in combinations(sanger_seqs, 2):
    a = calcSeqAgreement(combo[0],combo[1])
    scores.append(a[2])
print np.mean(scores)
print np.std(scores)

scores = []
for combo in combinations(illumina_seqs, 2):
    a = calcSeqAgreement(combo[0],combo[1])
    scores.append(a[2])
print np.mean(scores)
print np.std(scores)
